Replace `SoundPool` implementation with `MediaPlayer`
The `MediaPlayer` implementation replaces the `SoundPool` implementation because it exposes certain properties required by the `kha.audio1.AudioChannel` interface.

When a `kha.Sound` asset is created, it is passed an `AssetFileDescriptor`. When the asset is to be handled by `kha.audio2.Audio`, it is passed on to `kha.android.audio1.AndroidAudioChannel.ChannelHandler`. This handles the creation of `AndroidAudioChannel` instances, which it returns back to `Audio`. `Audio` then sets the channel up and returns it.

Every channel is a wrapper around `MediaPlayer`. When a channel is stopped, the `MediaPlayer` is released by the `ChannelHandler`, and space for new channels is opened.